+++ pyabsa/core/tc/classic/__glove__/dataset_utils/data_utils_for_training.py
@@ -17,7 +17,7 @@ def build_tokenizer(dataset_list, max_seq_len, dat_fname, opt):
     if not os.path.exists('run/{}'.format(dataset_name)):
         os.makedirs('run/{}'.format(dataset_name))
     tokenizer_path = 'run/{}/{}'.format(dataset_name, dat_fname)
-    if os.path.exists(tokenizer_path):
+    if os.path.exists(tokenizer_path) and not opt.overwrite_cache:
         print('Loading tokenizer on {}'.format(tokenizer_path))
         tokenizer = pickle.load(open(tokenizer_path, 'rb'))
     else:
@@ -32,7 +32,8 @@ def build_tokenizer(dataset_list, max_seq_len, dat_fname, opt):
 
         tokenizer = Tokenizer(max_seq_len)
         tokenizer.fit_on_text(text)
-        pickle.dump(tokenizer, open(tokenizer_path, 'wb'))
+        if opt.overwrite_cache:
+            pickle.dump(tokenizer, open(tokenizer_path, 'wb'))
     return tokenizer
 
 
@@ -58,8 +59,7 @@ class Tokenizer(object):
         if self.lower:
             text = text.lower()
         words = text.split()
-        unknownidx = len(self.word2idx) + 1
-        sequence = [self.word2idx[w] if w in self.word2idx else unknownidx for w in words]
+        sequence = [self.word2idx[w] if w in self.word2idx else 0 for w in words]
         if len(sequence) == 0:
             sequence = [0]
         if reverse:
